+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## v1.3
+
+### Changed
+- Update of the 'aggColumnUsername' function 
+
+### Added:
+- Added "deleteEmbed" function that deletes all tweets from all columns 
+
 ## v1.2
 
 ### New Features 
@@ -9,10 +17,10 @@
 ### Changed
 - Update of the 'README.md' file 
 
-### Feature Change **
+### Feature Change
 - Taking into account the number of columns in the page Notion in a dynamic way
 
-### Added: **
+### Added:
 - Added "checkTwitterUsername" function that checks if a Twitter user exists 
 - Added "getUsernames" function that retrieves the username in the column title of the Notion page 
 - Added function "aggColumnUsername" which calls sub-functions that check that the username exists on twitter and that the title of the column is of type "heading_1